move helper to library
